Merge from master repo

+++ deps/hiredis/sockcompat.h
@@ -49,9 +49,10 @@
 #include <winsock2.h>
 #include <ws2tcpip.h>
 #include <stddef.h>
+#include <errno.h>
 
 #ifdef _MSC_VER
-typedef signed long ssize_t;
+typedef long long ssize_t;
 #endif
 
 /* Emulate the parts of the BSD socket API that we need (override the winsock signatures). */
